* javafx: (170 commits)
  Introduces a placeholder for a DeviceView (a UI representing current data from a device)
  Implements the polygon drawing in the data visualization app
  JavaFX canvas affine transforms use rotation in degrees
  Removes some stray output to the console
  By default don't yet mark devices as "disconnected" in case the device is not serially connected and no DeviceConnectivity will ever arrive
  Working toward getting the visualization app running again using a JavaFX canvas
  Adds back "required" property to vital and bind ignoreZero properly to the UI
  Cleans up the SimControl window a little bit
  updated DeviceClock to use java.time.Instant
  Rewires the functionality of the Data Collector App to start/stop persisters
  Bind device node labels to makeAndModel; hide stacked persister UIs by default.
  Mustn't pass null for time any longer.
  Will use whatever properly configured (update-alternatives) JVM is on the beaglebone adapter from now on.
  Turns out the JavaFX App thread is statically initialized as a side effect of loading the javafx.application.Application class so for 'command line mode' I had to take some pains not to even reference that class.
  Adds a tooltip with IP address
  Restore the "OpenICE" title when going "back" from an app
  first stab at the clock refactor
  more work in progress.  fixed some property binding messes. use a common NumericInstanceModel throughout, etc.
  Important fix to include get_key_value before notifying that an instance is no longer alive
  PB840 field mappings are updated. Added extra logging to PB840 driver. Increased the SENDF time from 1 to 2 seconds.
  ...

Conflicts:
	interop-lab/demo-apps/src/main/java/org/mdpnp/apps/testapp/Configuration.java
	interop-lab/demo-apps/src/main/java/org/mdpnp/apps/testapp/DeviceFactory.java

+++ interop-lab/demo-apps/src/main/java/org/mdpnp/apps/testapp/pca/VitalSign.java
@@ -23,15 +23,16 @@ import org.mdpnp.apps.testapp.vital.VitalModel;
  */
 public enum VitalSign {
     HeartRate("Heart Rate", "bpm", new String[] { rosetta.MDC_PULS_OXIM_PULS_RATE.VALUE, rosetta.MDC_PULS_RATE_NON_INV.VALUE,
-            rosetta.MDC_BLD_PULS_RATE_INV.VALUE, rosetta.MDC_ECG_CARD_BEAT_RATE.VALUE, rosetta.MDC_ECG_HEART_RATE.VALUE }, 40f, 140f, 20f, 160f, 0f, 250f, 5000L, 5000L, Color.green), SpO2(
-            "SpO\u2082", "%", new String[] { rosetta.MDC_PULS_OXIM_SAT_O2.VALUE }, 95f, 100f, 85f, 100f, 50f, 100f, 5000L, 5000L, Color.pink), RespiratoryRate(
-            "Resp Rate", "bpm", new String[] { rosetta.MDC_RESP_RATE.VALUE, rosetta.MDC_TTHOR_RESP_RATE.VALUE, rosetta.MDC_CO2_RESP_RATE.VALUE, rosetta.MDC_VENT_RESP_RATE.VALUE, rosetta.MDC_AWAY_RESP_RATE.VALUE }, 10f, 18f, 4f, 35f, 0f, 40f, 5000L, 5000L, Color.yellow), EndTidalCO2(
-            "etCO\u2082", "mmHg", new String[] { rosetta.MDC_AWAY_CO2_ET.VALUE }, 20f, 45f, 10f, 75f, 0f, 120f, 5000L, 5000L, Color.yellow), Temperature(
-            "Temp", "\u00B0C", new String[] { rosetta.MDC_TEMP_BLD.VALUE }, 35f, 39f, 32f, 42f, 20f, 60f, 5000L, 5000L, Color.black);
+            rosetta.MDC_BLD_PULS_RATE_INV.VALUE, rosetta.MDC_ECG_CARD_BEAT_RATE.VALUE, rosetta.MDC_ECG_HEART_RATE.VALUE }, 40.0, 140.0, 20.0, 160.0, 0.0, 250.0, 5000L, 5000L, Color.green), 
+            SpO2(
+            "SpO\u2082", "%", new String[] { rosetta.MDC_PULS_OXIM_SAT_O2.VALUE }, 95.0, 100.0, 85.0, 100.0, 50.0, 100.0, 5000L, 5000L, Color.pink), RespiratoryRate(
+            "Resp Rate", "bpm", new String[] { rosetta.MDC_RESP_RATE.VALUE, rosetta.MDC_TTHOR_RESP_RATE.VALUE, rosetta.MDC_CO2_RESP_RATE.VALUE, rosetta.MDC_VENT_RESP_RATE.VALUE, rosetta.MDC_AWAY_RESP_RATE.VALUE }, 10.0, 18.0, 4.0, 35.0, 0.0, 40.0, 5000L, 5000L, Color.yellow), EndTidalCO2(
+            "etCO\u2082", "mmHg", new String[] { rosetta.MDC_AWAY_CO2_ET.VALUE }, 20.0, 45.0, 10.0, 75.0, 0.0, 120.0, 5000L, 5000L, Color.yellow), Temperature(
+            "Temp", "\u00B0C", new String[] { rosetta.MDC_TEMP_BLD.VALUE }, 35.0, 39.0, 32.0, 42.0, 20.0, 60.0, 5000L, 5000L, Color.black);
     ;
 
-    VitalSign(String label, String units, String[] metric_ids, Float startingLow, Float startingHigh, Float criticalLow, Float criticalHigh,
-            float minimum, float maximum, Long valueMsWarningLow, Long valueMsWarningHigh, Color color) {
+    VitalSign(String label, String units, String[] metric_ids, Double startingLow, Double startingHigh, Double criticalLow, Double criticalHigh,
+            double minimum, double maximum, Long valueMsWarningLow, Long valueMsWarningHigh, Color color) {
         this.label = label;
         this.units = units;
         this.metric_ids = metric_ids;
@@ -53,8 +54,8 @@ public enum VitalSign {
 
     private final String label, units;;
     private final String[] metric_ids;
-    private final Float startingLow, startingHigh, criticalLow, criticalHigh;
+    private final Double startingLow, startingHigh, criticalLow, criticalHigh;
     private final Long valueMsWarningLow, valueMsWarningHigh;
-    private final float minimum, maximum;
+    private final double minimum, maximum;
     private final Color color;
 }
